requirements.txt
# Python >=3.11,<3.13
fastapi==0.110.2
uvicorn==0.29.0
sqlmodel==0.0.16
pydantic==2.7.1
python-dotenv==1.0.1
openai==1.30.0
